requirements.txt
mcp>=1.2.0
python-dotenv>=1.0.0
requests>=2.32.0